ENH: Update vtkMRMLScalarVolumeDisplayNode pipeline as soon as properties get modified
not just when UpdateImageDataPipeline is called.

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=15324
git-svn-id: http://svn.slicer.org/Slicer4/trunk@15324 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/EMSegment/MRML/vtkMRMLEMSVolumeCollectionNode.cxx
@@ -227,7 +227,7 @@ int
 vtkMRMLEMSVolumeCollectionNode::
 GetNumberOfVolumes() const
 {
-  return this->KeyList.size();
+  return static_cast<int>(this->KeyList.size());
 }
 
 void
